+++ magda-web-client/src/actions/userManagementActions.ts
@@ -4,7 +4,6 @@ import { actionTypes } from "../constants/ActionTypes";
 import { Dispatch, GetState } from "../types";
 import { FacetAction } from "../helpers/datasetSearch";
 import { fetchContent } from "actions/contentActions";
-import request from "helpers/request";
 
 export function requestWhoAmI() {
     return (dispatch: Dispatch, getState: GetState) => {
@@ -151,17 +150,4 @@ export function receiveAuthProvidersError(err: any): any {
         type: actionTypes.RECEIVE_AUTH_PROVIDERS_ERROR,
         err
     };
-}
-
-export function listUsers() {
-    return request("GET", config.authApiUrl + "users/all");
-}
-
-export function updateUser(user, patch) {
-    return request(
-        "PUT",
-        config.authApiUrl + "users/" + user,
-        patch,
-        "application/json"
-    );
 }