+++ magda-web-client/src/helpers/getRequestNoCache.ts
@@ -1,0 +1,16 @@
+import request from "./request";
+import createNoCacheFetchOptions from "api-clients/createNoCacheFetchOptions";
+
+export default async function getRequestNoCache<T = any, CT = string>(
+    url: string,
+    extraFetchOptions: RequestInit = {}
+) {
+    return await request<T, CT>(
+        "get",
+        url,
+        undefined,
+        undefined,
+        undefined,
+        createNoCacheFetchOptions(extraFetchOptions)
+    );
+}